libfreerdp-core: fix potential overflow in parsing of fragmented RemoteFX packets

+++ libfreerdp-core/surface.c
@@ -58,11 +58,11 @@ static int update_recv_surfcmd_frame_marker(rdpUpdate* update, STREAM* s)
 	return 6;
 }
 
-boolean update_recv_surfcmds(rdpUpdate* update, uint16 size, STREAM* s)
+boolean update_recv_surfcmds(rdpUpdate* update, uint32 size, STREAM* s)
 {
 	uint8* mark;
 	uint16 cmdType;
-	uint16 cmdLength;
+	uint32 cmdLength;
 
 	while (size > 2)
 	{
